hack job to make testsuite executable on non-NES 6502 cpu emulators
in order to use this, disable BCD in your 6502 emulator so it behaves
like the Ricoh 2A03 (which is an exact 6502 copy with 5 transistors
dealing with the patented BCD circuit scratched off[0]), and if you
want output, add some code that catches the multi-byte nop opcode 0xfc
followed by 0x13, 0x37, which i repurposed as "syscall" opcode to signal
the emulator that a byte to output to stdout waits at 0x2000.

the entrypoint is at 0x8004, but you can also just jump to the reset vector.

[0]: http://web.archive.org/web/20210405071434/http://www.visual6502.org/wiki/index.php?title=6502DecimalMode#Decimal_mode_and_the_NES.27_RP2A03G